refactor(mfrac): extract LineThickness value object into mfrac folder
Move mfrac into its own folder and split linethickness parsing out of the
converter into a dedicated LineThickness value object, so MFrac orchestrates
fraction forms while thickness interpretation is unit-tested in isolation.
