+++ src/data/usecases/mathml-to-latex-convertion/converters/mfrac/index.ts
@@ -1,0 +1,1 @@
+export { MFrac } from './mfrac';
